+++ components/job-orchestration/job_orchestration/scheduler/search/search_scheduler.py
@@ -16,7 +16,8 @@ import pymongo
 from clp_py_utils.clp_config import CLP_METADATA_TABLE_PREFIX, CLPConfig, SEARCH_JOBS_TABLE_NAME
 from clp_py_utils.clp_logging import get_logger, get_logging_formatter, set_logging_level
 from clp_py_utils.core import read_yaml_config_file
-from clp_py_utils.sql_adapter import exception_default_value, SQL_Adapter
+from clp_py_utils.decorators import exception_default_value
+from clp_py_utils.sql_adapter import SQL_Adapter
 from job_orchestration.executor.search.fs_search_task import search
 from job_orchestration.scheduler.constants import SearchJobStatus
 from job_orchestration.scheduler.job_config import SearchConfig
