+++ components/job-orchestration/pyproject.toml
@@ -22,7 +22,6 @@ pydantic = "^1.10.13"
 pymongo = "^4.6.1"
 PyYAML = "^6.0.1"
 redis = "^5.0.1"
-sqlalchemy = "~2.0"
 
 [build-system]
 requires = ["poetry-core"]
